chore: make cargo clippy happy

+++ src/magic_mount/try_umount.rs
@@ -82,7 +82,7 @@ where
         let ret = libc::ioctl(fd as libc::c_int, KSU_IOCTL_ADD_TRY_UMOUNT as u64, &cmd);
 
         #[cfg(not(target_env = "gnu"))]
-        let ret = libc::ioctl(fd as libc::c_int, KSU_IOCTL_ADD_TRY_UMOUNT as i32, &cmd);
+        let ret = libc::ioctl(fd as libc::c_int, KSU_IOCTL_ADD_TRY_UMOUNT, &cmd);
 
         if ret < 0 {
             log::error!(
